scene: fix window display jitter
when firefox local content is refreshed after window maximization
firefox's window display jitter.  The reason for the jitter is the
inconsistent rendering attributes of the texture before and after.
When window maimization the target scale is 1, after that the scale
is the output scale. traget scale isn't same, so the attributes is
different.